Reorganize BigInteger
