+++ scicrypt-bigint/src/arithmetic/sub.rs
@@ -1,0 +1,9 @@
+use std::ops::SubAssign;
+
+use crate::BigInteger;
+
+impl SubAssign<&BigInteger> for BigInteger {
+    fn sub_assign(&mut self, rhs: &BigInteger) {
+        todo!()
+    }
+}
