+++ scicrypt-bigint/src/lib.rs
@@ -1,6 +1,10 @@
 #![feature(int_roundings)]
 #![feature(test)]
-use std::{ops::{AddAssign, Mul, RemAssign, Rem, Add, ShrAssign, Shr, DivAssign, MulAssign, SubAssign}, cmp::{max, min}, mem::MaybeUninit, ffi::{CString, CStr}, fmt::{Display, Debug}, ptr::null_mut, alloc::Layout};
+mod arithmetic;
+mod binary;
+mod modular;
+
+use std::{cmp::min, mem::MaybeUninit, ffi::{CString, CStr}, fmt::{Display, Debug}, ptr::null_mut};
 
 use gmp_mpfr_sys::gmp::{mpz_t, self};
 
@@ -154,213 +158,6 @@ impl BigInteger {
         self.value.size == 0
     }
 
-    /// Compute `self` to the power `exponent` modulo an odd `modulus`. The computation takes time that scales with the specified size of the `exponent` and `modulus`.
-    pub fn pow_mod(&self, exponent: &BigInteger, modulus: &BigInteger) -> BigInteger {
-        debug_assert!(!self.is_zero(), "the base must be larger than 0");
-        debug_assert!(!modulus.is_zero(), "the modulus must be larger than 0");
-        // TODO: debug_assert!() that the modulus is ODD
-        // TODO: debug_assert!() that the exponent's bitsize is smaller than its size_in_bits
-        debug_assert!(exponent.size_in_bits > 0, "the exponent must be larger than 0");
-
-        debug_assert_eq!(modulus.size_in_bits as i32, modulus.value.size * GMP_NUMB_BITS as i32, "the modulus' size in bits must be tight with its actual size");
-
-        // TODO: Probably we should also assert that the modulus does not contain less limbs than the other operands
-
-        let mut result = BigInteger::init(modulus.value.size);
-
-        let enb = exponent.size_in_bits as u64;
-
-        unsafe {
-            let scratch_size =
-                gmp::mpn_sec_powm_itch(self.value.size as i64, enb, modulus.value.size as i64) as usize
-                    * GMP_NUMB_BITS as usize
-                    / 8;
-
-            if scratch_size == 0 {
-                gmp::mpn_sec_powm(
-                    result.value.d.as_mut(),
-                    self.value.d.as_ptr(),
-                    self.value.size as i64,
-                    exponent.value.d.as_ptr(),
-                    enb,
-                    modulus.value.d.as_ptr(),
-                    modulus.value.size as i64,
-                    null_mut(),
-                );
-
-                result.value.size = modulus.value.size;
-                return result;
-            }
-
-            let scratch_layout = Layout::from_size_align(scratch_size, ALIGN).unwrap();
-            let scratch = std::alloc::alloc(scratch_layout);
-
-            gmp::mpn_sec_powm(
-                result.value.d.as_mut(),
-                self.value.d.as_ptr(),
-                self.value.size as i64,
-                exponent.value.d.as_ptr(),
-                enb,
-                modulus.value.d.as_ptr(),
-                modulus.value.size as i64,
-                scratch as *mut u64,
-            );
-
-            std::alloc::dealloc(scratch, scratch_layout);
-
-            result.value.size = modulus.value.size;
-            result
-        }
-    }
-
-    /// Computes `self^-1 mod modulus`, taking ownership of `self`. Returns None if no inverse exists. `modulus` must be odd.
-    pub fn invert(self, modulus: &BigInteger) -> Option<BigInteger> {
-        // TODO: Verify that the input must be smaller than the modulus (is this indeed true?)
-        //assert_eq!(self.supposed_size, modulus.supposed_size);
-        //self.supposed_size = modulus.inner.size as i64;
-
-        debug_assert_eq!(modulus.size_in_bits.div_ceil(GMP_NUMB_BITS as i64) as i32, modulus.value.size, "the modulus' size in bits must match its actual size");
-        //debug_assert_eq!(modulus.size_in_bits as i32, modulus.value.size * GMP_NUMB_BITS as i32, "the modulus' size in bits must be tight with its actual size");
-        debug_assert_eq!(modulus.size_in_bits, self.size_in_bits, "the modulus must have the same size as self");
-
-        // FIXME: This is not constant-time
-        //debug_assert_eq!(modulus.value.size, self.value.size, "the modulus must have the same actual size as self");
-        // if self.value.size != modulus.value.size {
-        //     self += modulus;
-        // }
-        debug_assert_eq!(modulus.value.size, self.value.size, "the modulus must have the same actual size as self");
-
-        //self += modulus;
-
-        let mut result = BigInteger::init(modulus.value.size);
-
-        unsafe {
-            let scratch_size = gmp::mpn_sec_invert_itch(modulus.value.size as i64)
-                as usize
-                * GMP_NUMB_BITS as usize
-                / 8;
-
-            if scratch_size == 0 {
-                let is_valid = gmp::mpn_sec_invert(
-                    result.value.d.as_mut(),
-                    self.value.d.as_ptr(),
-                    modulus.value.d.as_ptr(),
-                    modulus.value.size as i64,
-                    (self.size_in_bits + modulus.size_in_bits) as u64,
-                    null_mut(),
-                );
-
-                // Check if an inverse exists
-                if is_valid == 0 {
-                    return None;
-                }
-
-                result.value.size = modulus.value.size;
-                result.size_in_bits = modulus.size_in_bits;
-                return Some(result);
-            }
-
-            let scratch_layout = Layout::from_size_align(scratch_size, ALIGN).unwrap();
-            let scratch = std::alloc::alloc(scratch_layout);
-
-            let is_valid = gmp::mpn_sec_invert(
-                result.value.d.as_mut(),
-                self.value.d.as_ptr(),
-                modulus.value.d.as_ptr(),
-                modulus.value.size as i64,
-                (self.size_in_bits + modulus.size_in_bits) as u64,
-                scratch as *mut u64,
-            );
-
-            std::alloc::dealloc(scratch, scratch_layout);
-
-            // Check if an inverse exists
-            if is_valid == 0 {
-                return None;
-            }
-
-            result.value.size = modulus.value.size;
-            result.size_in_bits = modulus.size_in_bits;
-            return Some(result);
-        }
-    }
-
-    // /// Computes `self^-1 mod modulus`, taking ownership of `self`. Returns None if no inverse exists. `modulus` must be odd.
-    // pub fn invert(mut self, modulus: &BigInteger) -> Option<BigInteger> {
-    //     //assert_eq!(self.supposed_size, modulus.supposed_size);
-    //     //self.supposed_size = modulus.inner.size as i64;
-
-    //     debug_assert_eq!(modulus.size_in_bits.div_ceil(GMP_NUMB_BITS as i64) as i32, modulus.value.size, "the modulus' size in bits must match its actual size");
-    //     //debug_assert_eq!(modulus.size_in_bits as i32, modulus.value.size * GMP_NUMB_BITS as i32, "the modulus' size in bits must be tight with its actual size");
-    //     debug_assert_eq!(modulus.size_in_bits, self.size_in_bits, "the modulus must have the same size as self");
-
-    //     dbg!(&self);
-    //     dbg!(&modulus);
-    //     // FIXME: This is not constant-time
-    //     //debug_assert_eq!(modulus.value.size, self.value.size, "the modulus must have the same actual size as self");
-    //     if self.value.size != modulus.value.size {
-    //         self += modulus;
-    //     }
-    //     debug_assert_eq!(modulus.value.size, self.value.size, "the modulus must have the same actual size as self");
-
-    //     //self += modulus;
-
-    //     //let mut result = BigInteger::init(modulus.value.size);
-
-    //     unsafe {
-    //         let scratch_size = gmp::mpn_sec_invert_itch(modulus.value.size as i64)
-    //             as usize
-    //             * GMP_NUMB_BITS as usize
-    //             / 8;
-    //         dbg!(scratch_size);
-
-    //         if scratch_size == 0 {
-    //             let is_valid = gmp::mpn_sec_invert(
-    //                 self.value.d.as_mut(),
-    //                 self.value.d.as_ptr(),
-    //                 modulus.value.d.as_ptr(),
-    //                 modulus.value.size as i64,
-    //                 (self.size_in_bits + modulus.size_in_bits) as u64,
-    //                 null_mut(),
-    //             );
-
-    //             // Check if an inverse exists
-    //             if is_valid == 0 {
-    //                 return None;
-    //             }
-
-    //             // TODO: This can probably be removed
-    //             self.value.size = modulus.value.size;
-    //             self.size_in_bits = modulus.size_in_bits;
-    //             return Some(self);
-    //         }
-
-    //         let scratch_layout = Layout::from_size_align(scratch_size, ALIGN).unwrap();
-    //         let scratch = std::alloc::alloc(scratch_layout);
-
-    //         let is_valid = gmp::mpn_sec_invert(
-    //             self.value.d.as_mut(),
-    //             self.value.d.as_ptr(),
-    //             modulus.value.d.as_ptr(),
-    //             modulus.value.size as i64,
-    //             (self.size_in_bits + modulus.size_in_bits) as u64,
-    //             scratch as *mut u64,
-    //         );
-
-    //         std::alloc::dealloc(scratch, scratch_layout);
-
-    //         // Check if an inverse exists
-    //         if is_valid == 0 {
-    //             return None;
-    //         }
-
-    //         // TODO: This can probably be removed
-    //         self.value.size = modulus.value.size;
-    //         self.size_in_bits = modulus.size_in_bits;
-    //         return Some(self);
-    //     }
-    // }
-
     // Computes the least common multiple between self and other. This function is not constant-time.
     pub fn lcm(&self, other: &BigInteger) -> BigInteger {
         let mut result = BigInteger::init(self.value.size);
@@ -370,113 +167,6 @@ impl BigInteger {
         }
 
         result
-    }
-
-    pub fn square(&self) -> BigInteger {
-        // TODO: Switch to more efficient squaring function
-        self * self
-    }
-}
-
-impl AddAssign<&BigInteger> for BigInteger {
-    fn add_assign(&mut self, rhs: &Self) {
-        let n = min(self.value.size, rhs.value.size);
-
-        if n == 0 {
-            return;
-        }
-
-        unsafe {
-            let carry = gmp::mpn_add_n(
-                self.value.d.as_mut(),
-                self.value.d.as_ptr(),
-                rhs.value.d.as_ptr(),
-                n as i64,
-            );
-
-            let largest_size = max(self.value.size, rhs.value.size) as i32;
-
-            self.value.size = largest_size + carry as i32;
-            self.size_in_bits = max(self.size_in_bits, rhs.size_in_bits) + carry as i64;
-        }
-    }
-}
-
-impl Add<&BigInteger> for BigInteger {
-    type Output = BigInteger;
-
-    fn add(mut self, rhs: &Self) -> Self::Output {
-        self += rhs;
-        self
-    }
-}
-
-// TODO: Also implement addition with u64 using `mpn_sec_add_1`
-
-impl Mul for &BigInteger {
-    type Output = BigInteger;
-
-    fn mul(self, rhs: Self) -> Self::Output {
-        if rhs.value.size > self.value.size {
-            return rhs * self;
-        }
-
-        debug_assert_eq!(self.size_in_bits.div_ceil(GMP_NUMB_BITS as i64) as i32, self.value.size, "the operands' size in bits must match their actual size");
-        debug_assert_eq!(rhs.size_in_bits.div_ceil(GMP_NUMB_BITS as i64) as i32, rhs.value.size, "the operands' size in bits must match their actual size");
-
-        let mut result = BigInteger::init(self.value.size + rhs.value.size);
-
-        unsafe {
-            let scratch_size = gmp::mpn_sec_mul_itch(self.value.size as i64, rhs.value.size as i64)
-                as usize
-                * GMP_NUMB_BITS as usize
-                / 8;
-
-            if scratch_size == 0 {
-                gmp::mpn_sec_mul(
-                    result.value.d.as_mut(),
-                    self.value.d.as_ptr(),
-                    self.value.size as i64,
-                    rhs.value.d.as_ptr(),
-                    rhs.value.size as i64,
-                    null_mut(),
-                );
-
-                result.value.size = self.value.size + rhs.value.size;
-                result.size_in_bits = self.size_in_bits + rhs.size_in_bits;
-                return result;
-            }
-
-            let scratch_layout = Layout::from_size_align(scratch_size, ALIGN).unwrap();
-            let scratch = std::alloc::alloc(scratch_layout);
-
-            gmp::mpn_sec_mul(
-                result.value.d.as_mut(),
-                self.value.d.as_ptr(),
-                self.value.size as i64,
-                rhs.value.d.as_ptr(),
-                rhs.value.size as i64,
-                scratch as *mut u64,
-            );
-
-            std::alloc::dealloc(scratch, scratch_layout);
-
-            result.value.size = self.value.size + rhs.value.size;
-            result.size_in_bits = self.size_in_bits + rhs.size_in_bits;
-            result
-        }
-    }
-}
-
-impl DivAssign<&BigInteger> for BigInteger {
-    fn div_assign(&mut self, rhs: &BigInteger) {
-        todo!()
-    }
-}
-
-impl SubAssign<&BigInteger> for BigInteger {
-    fn sub_assign(&mut self, rhs: &BigInteger) {
-        todo!()
     }
 }
 
@@ -489,93 +179,7 @@ impl PartialEq for BigInteger {
     }
 }
 
-impl Eq for BigInteger {
-    
-}
-
-impl RemAssign<&BigInteger> for BigInteger {
-    fn rem_assign(&mut self, rhs: &Self) {
-        debug_assert_eq!(self.size_in_bits.div_ceil(GMP_NUMB_BITS as i64) as i32, self.value.size, "the operands' size in bits must match their actual size");
-        debug_assert_eq!(rhs.size_in_bits.div_ceil(GMP_NUMB_BITS as i64) as i32, rhs.value.size, "the operands' size in bits must match their actual size");
-
-        unsafe {
-            let scratch_size = gmp::mpn_sec_div_r_itch(self.value.size as i64, rhs.value.size as i64)
-                as usize
-                * GMP_NUMB_BITS as usize
-                / 8;
-
-            if scratch_size == 0 {
-                gmp::mpn_sec_div_r(
-                    self.value.d.as_mut(),
-                    self.value.size as i64,
-                    rhs.value.d.as_ptr(),
-                    rhs.value.size as i64,
-                    null_mut(),
-                );
-
-                self.value.size = rhs.value.size;
-                self.size_in_bits = rhs.size_in_bits;
-                return;
-            }
-
-            let scratch_layout = Layout::from_size_align(scratch_size, ALIGN).unwrap();
-            let scratch = std::alloc::alloc(scratch_layout);
-
-            gmp::mpn_sec_div_r(
-                self.value.d.as_mut(),
-                self.value.size as i64,
-                rhs.value.d.as_ptr(),
-                rhs.value.size as i64,
-                scratch as *mut u64,
-            );
-
-            std::alloc::dealloc(scratch, scratch_layout);
-
-            self.value.size = rhs.value.size;
-            self.size_in_bits = rhs.size_in_bits;
-        }
-    }
-}
-
-impl Rem<&BigInteger> for BigInteger {
-    type Output = BigInteger;
-
-    fn rem(mut self, rhs: &BigInteger) -> Self::Output {
-        self %= rhs;
-        self
-    }
-}
-
-/// Not a constant-time function: Reveals the actual size of self.
-impl ShrAssign<u32> for BigInteger {
-    fn shr_assign(&mut self, rhs: u32) {
-        assert!(1 <= rhs);
-        assert!(rhs as u64 <= GMP_NUMB_BITS - 1);
-
-        unsafe {
-            gmp::mpn_rshift(self.value.d.as_mut(), self.value.d.as_ptr(), self.value.size as i64, rhs);
-        }
-    }
-}
-
-/// Not a constant-time function: Reveals the actual size of self.
-impl Shr<u32> for &BigInteger {
-    type Output = BigInteger;
-
-    fn shr(self, rhs: u32) -> Self::Output {
-        assert!(1 <= rhs);
-        assert!(rhs as u64 <= GMP_NUMB_BITS - 1);
-
-        let mut result = BigInteger::init(self.value.size);
-
-        unsafe {
-            gmp::mpn_rshift(result.value.d.as_mut(), self.value.d.as_ptr(), self.value.size as i64, rhs);
-        }
-
-        result.value.size = self.value.size;
-        result
-    }
-}
+impl Eq for BigInteger { }
 
 impl Clone for BigInteger {
     fn clone(&self) -> Self {
